fixed authentication errors, common bug fixes

+++ src/components/Snippet/index.tsx
@@ -3,6 +3,7 @@ import React from "react";
 
 import { toast } from "react-toastify";
 import { FiShare } from "react-icons/fi";
+import { AiOutlineEye } from "react-icons/ai";
 import { MdOutlineDeleteForever, MdOutlineEdit } from "react-icons/md";
 
 import dayjs from "@/utility/dayjs";
@@ -18,7 +19,6 @@ export default function Snippet({
   created_at,
   updated_at,
 }: ISnippet) {
-
   const user = useAuthStore((state) => state.user);
   const { setMode, setVisible, setSelectedSnippetId } = useModalStore(
     (state) => state
@@ -39,6 +39,11 @@ export default function Snippet({
     }
   };
 
+  const onViewClicked = () => {
+    // setMode("view");
+    // setVisible(true);
+  };
+
   const onEditClicked = () => {
     setMode("edit");
     setSelectedSnippetId(id);
@@ -47,6 +52,12 @@ export default function Snippet({
 
   const onDeleteClicked = async () => {
     if (!user) return;
+
+    const answer = window.confirm(
+      "Are you sure you want to delete this snippet?"
+    );
+    if (!answer) return;
+
     const res = await deleteSnippet(id);
     if (res) {
       toast("Snippet Deleted", {
@@ -80,12 +91,21 @@ export default function Snippet({
           </div>
           <div className="flex flex-row space-x-4">
             <div
+              title="View Snippet"
+              onClick={onViewClicked}
+              className="cursor-pointer rounded-full bg-[#e2ffeb] p-2"
+            >
+              <AiOutlineEye color="#32f56a" size={20} />
+            </div>
+            <div
+              title="Edit Snippet"
               onClick={onEditClicked}
               className="cursor-pointer rounded-full bg-[#e2ebff] p-2"
             >
               <MdOutlineEdit color="#325ddd" size={20} />
             </div>
             <div
+              title="Delete Snippet"
               onClick={onDeleteClicked}
               className="cursor-pointer rounded-full bg-[#ffe2e2] p-2"
             >
@@ -101,12 +121,12 @@ export default function Snippet({
           <div className="flex flex-row items-center space-x-6">
             <span className="text-xs font-light text-slate-400">
               {/* @ts-ignore */}
-              created {dayjs(created_at.toDate()).fromNow()}
+              created {dayjs(created_at?.toDate()).fromNow()}
             </span>
             {updated_at && (
               <span className="text-xs font-light text-slate-400">
                 {/* @ts-ignore */}• updated{" "}
-                {dayjs(updated_at.toDate()).fromNow()}
+                {dayjs(updated_at?.toDate()).fromNow()}
               </span>
             )}
           </div>
